dashboard/agent-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cua-agent"
version = "0.1.0"
description = "Computer Use Agent service for vLLM Semantic Router Dashboard"
authors = [{ name = "vLLM Team" }]
keywords = ["fastapi", "api", "backend", "automation", "e2b", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
    "smolagents[openai,litellm]>=1.20.0",
    "e2b-desktop>=2.0.0",
    "huggingface_hub>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/cua_agent"]

[tool.hatch.build.targets.sdist]
include = ["/src"]
